fix ARM-M class definition
without the proper parameter for Elf.Abi, `reset_architecture` produces a
warning. This warnin is not only inconvenient, but does stop the debugger from
attaching automatically.

according to i.e. https://community.arm.com/support-forums/f/architectures-and-processors-forum/51867/how-can-i-know-if-an-elf-file-is-for-cortex-a-or-cortex-m
there is no different ABI code form ARM platforms. CPU Type would be needed to
make a proper distinction.