Add support for BlackMagicProbe (#99)
This adds three things:

`gef-bmp-remote` command, to connect to a black magic probe. This is its
own command because it connects to a tty instead of a host/port.

`GefBMPRemoteSessionManager`, which handles connecting and such.

`ARMBlackMagicProbe`, a new arch, which handles its own memory
management and has different register names that e.g. openocd.