Implement "got-audit" command. (#112)
## Description

The "got-audit" command examines the symbols in the GOT, and prints a
list of symbols and the path of the file that provides the mapped memory
that the value points to. Additionally, it will print errors if a symbol
is provided by multiple shared libraries, or if a symbol in the GOT
points to a library that doesn't provide it.

To prevent future attacks similar to the liblzma attack, I'd like to
implement a tool to audit the GOT of a running process which can be used
in Fedora (and elsewhere) test infrastructure to look for signs of
tampering.

## How Has This Been Tested ?

"Tested" indicates that the PR works *and* the unit test (i.e. `make
test`) run passes without issue.

*  [ ] x86-32
*  [x] x86-64
*  [ ] ARM
*  [ ] AARCH64
*  [ ] MIPS
*  [ ] POWERPC
*  [ ] SPARC
*  [ ] RISC-V